Add --semesters argument to reset_topics script
